+++ Civi/DataProcessor/Source/Contact/HouseholdSource.php
@@ -6,6 +6,9 @@
 
 namespace Civi\DataProcessor\Source\Contact;
 
+use Civi\DataProcessor\DataFlow\SqlDataFlow\AndClause;
+use Civi\DataProcessor\DataFlow\SqlDataFlow\OrClause;
+use Civi\DataProcessor\DataFlow\SqlDataFlow\SimpleWhereClause;
 use Civi\DataProcessor\DataSpecification\DataSpecification;
 use Civi\DataProcessor\Source\AbstractCivicrmEntitySource;
 
@@ -121,6 +124,43 @@ class HouseholdSource extends AbstractCivicrmEntitySource {
   protected function addFilters($configuration) {
     parent::addFilters($configuration);
     $this->addFilter('contact_type', '=', 'Household');
+  }
+
+  /**
+   * Adds an inidvidual filter to the data source
+   *
+   * @param $filter_field_alias
+   * @param $op
+   * @param $values
+   *
+   * @throws \Exception
+   */
+  protected function addFilter($filter_field_alias, $op, $values) {
+    if ($filter_field_alias == 'contact_sub_type' && $op == 'IN') {
+      $contactTypeClauses = [];
+      foreach ($values as $value) {
+        $contactTypeSearchName = '%' . \CRM_Core_DAO::VALUE_SEPARATOR . $value . \CRM_Core_DAO::VALUE_SEPARATOR . '%';
+        $contactTypeClauses[] = new SimpleWhereClause($this->getSourceName(), 'contact_sub_type', 'LIKE', $contactTypeSearchName, 'String', TRUE);
+      }
+      if (count($contactTypeClauses)) {
+        $contactTypeClause = new OrClause($contactTypeClauses);
+        $entityDataFlow = $this->ensureEntity();
+        $entityDataFlow->addWhereClause($contactTypeClause);
+      }
+    } elseif ($filter_field_alias == 'contact_sub_type' && $op == 'NOT IN') {
+      $contactTypeClauses = [];
+      foreach($values as $value) {
+        $contactTypeSearchName = '%'.\CRM_Core_DAO::VALUE_SEPARATOR.$value.\CRM_Core_DAO::VALUE_SEPARATOR.'%';
+        $contactTypeClauses[] = new SimpleWhereClause($this->getSourceName(), 'contact_sub_type', 'NOT LIKE', $contactTypeSearchName, 'String',TRUE);
+      }
+      if (count($contactTypeClauses)) {
+        $contactTypeClause = new AndClause($contactTypeClauses);
+        $entityDataFlow = $this->ensureEntity();
+        $entityDataFlow->addWhereClause($contactTypeClause);
+      }
+    } else {
+      parent::addFilter($filter_field_alias, $op, $values);
+    }
   }